tests/requirements.txt
pytest
pytest-playwright
pytest-base-url
docker
